+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [2024-10-09] [Scaffold evals](https://github.com/RubricLab/memory/commit/3801514a795881c74ea225d02eeae001a07ee57a)
 - [2024-10-04] [bleed for mono](https://github.com/RubricLab/memory/commit/6db7d39072c60714068bcb00b07bbf917d76b4b8)
 - [2024-10-03] [mod lint script](https://github.com/RubricLab/memory/commit/e5a4392d2e7852f0d1424f730aa1316bfeb66f3b)
 - [2024-10-03] [gitignore turbo, format](https://github.com/RubricLab/memory/commit/fd8aedfe7144de98f9a405305e3b82839084fab7)